+++ edc-extensions/bpn-validation/bpn-validation-core/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -17,5 +17,6 @@
 #
 #  SPDX-License-Identifier: Apache-2.0
 #
-org.eclipse.tractusx.edc.validation.businesspartner.BusinessPartnerEvaluationExtension
+org.eclipse.tractusx.edc.validation.businesspartner.BusinessPartnerValidationExtension
 org.eclipse.tractusx.edc.validation.businesspartner.defaults.DefaultStoreProviderExtension
+org.eclipse.tractusx.edc.validation.businesspartner.LegacyBusinessPartnerValidationExtension
