+++ edc-extensions/bpn-validation/bpn-validation-spi/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -17,5 +17,5 @@
 #
 #  SPDX-License-Identifier: Apache-2.0
 #
-org.eclipse.tractusx.edc.validation.businesspartner.BusinessPartnerEvaluationExtension
+org.eclipse.tractusx.edc.validation.businesspartner.BusinessPartnerValidationExtension
 org.eclipse.tractusx.edc.validation.businesspartner.defaults.DefaultStoreProviderExtension
